requirements.txt
binstruct
bitstruct
chunker
construct
cstruct
destruct
EStruct
infi.instruct
netstruct
suitcase
vstruct
